requirements.txt
psycopg2
django-environ
urllib3
pycodestyle
